// src/lib/api/types/askResponseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */
import type { CitationDTO } from './citationDTO';
import type { UsageDTO } from './usageDTO';

export interface AskResponseBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  answer: string;
  /** @nullable */
  citations: CitationDTO[] | null;
  createdAt: string;
  latencyMs: number;
  model: string;
  requestId: string;
  sessionId: string;
  updatedAt: string;
  usage: UsageDTO;
}