// src/lib/api/types/conversationDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface ConversationDTO {
  accountId: string;
  createdAt: string;
  id: string;
  language: string;
  title: string;
  updatedAt: string;
}