// src/lib/api/types/deadEventDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface DeadEventDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  createdAt: string;
  errorMessage?: string;
  eventId: string;
  eventType: string;
  payload: string;
  replayCount: number;
  status: string;
}